refactor: extract reusable scam card component
