+++ app/lessons/all/page.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 // Corrected to use createBrowserClient for the 'use client' file
 import { createBrowserClient } from '@supabase/ssr';
 import { getScamCategory } from '../../utils/scam-helpers';
+import ScamCard from "@/components/ScamCard";
 
 interface Scam {
   id: number;
@@ -89,56 +90,12 @@ export default function BrowseKnowledge() {
             <p className="text-gray-400 font-bold">Scanning Registry...</p>
           </div>
         ) : scams.length > 0 ? (
-          scams.map((scam) => {
-            const category = getScamCategory(scam.tags);
-            
-            // Logic to handle titles for common scams
-            const displayTitle = isNaN(parseInt(scam.host[0])) 
-              ? scam.host.split('.')[0].toUpperCase() 
-              : "Unverified Source";
-
-            return (
-              /* FIXED: Changed bg-white to var(--card) and added border variables */
-              <div 
-                key={scam.id} 
-                className="group rounded-[2.5rem] p-10 shadow-[0_10px_40px_rgba(0,0,0,0.04)] border flex flex-col hover:shadow-2xl hover:-translate-y-1 transition-all duration-300"
-                style={{ background: 'var(--card)', borderColor: 'var(--card-border)' }}
-              >
-                <div className={`${category.bg} w-20 h-20 rounded-3xl flex items-center justify-center text-4xl mb-8 group-hover:scale-110 transition-transform`}>
-                  {category.icon}
-                </div>
-
-                <div className="flex items-center gap-2 mb-4">
-                  <span className={`text-[11px] font-black tracking-widest uppercase ${category.color}`}>
-                    {category.label}
-                  </span>
-                  <span className="text-gray-400 dark:text-gray-500 text-[11px] font-bold">• ACTIVE THREAT</span>
-                </div>
-
-                <h3 className="text-2xl font-black mb-4 leading-tight" style={{ color: 'var(--foreground)' }}>
-                  {displayTitle} Threat
-                </h3>
-
-                <p className="text-gray-400 dark:text-gray-400 text-base leading-relaxed mb-10 flex-grow font-medium">
-                  Our system flagged a <span className="text-gray-900 dark:text-gray-100 font-bold">{scam.threat.replace('_', ' ').toUpperCase()}</span> attempt from <strong>{scam.host}</strong>. 
-                  Always verify the sender before clicking.
-                </p>
-
-                <div className="flex items-center justify-between pt-8 border-t" style={{ borderColor: 'var(--card-border)' }}>
-                  <span className="text-[11px] font-black text-green-500 uppercase tracking-widest">
-                    Verified Risk
-                  </span>
-                  
-                  <Link 
-                    href={`/lessons/all/${scam.id}`}
-                    className="text-[#7042F4] font-black text-sm uppercase tracking-wider hover:opacity-70 flex items-center gap-2"
-                  >
-                    Safety Guide <span className="text-lg">→</span>
-                  </Link>
-                </div>
-              </div>
-            );
-          })
+          scams.map((scam) => (
+            <ScamCard
+              key={scam.id}
+              scam={scam}
+            />
+          ))
         ) : (
           /* FIXED: Updated empty state to use card variables */
           <div className="col-span-full text-center py-24 rounded-[3rem] border-2 border-dashed" style={{ background: 'var(--card)', borderColor: 'var(--card-border)' }}>
